Created displayContent module
Module to display the projects along with its lists was created. Functionality
added included creating a function that displays the project title in the
header of the section area. That function was attached to an event listener
so that when the project title in the sidebar area is click, the title is
displayed.