Refactored displayProject function
displayProject function was renamed to displayTasks and the code was
refactored. displayTasks is now the base code to pass task objects to. The
displayProject function now gets the tasks associated with the
selected project.

To support the change, additional functions will be added display all tasks,
weekly tasks, and current day.